refactor: refactor persons
